requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
google-api-core==2.5.0
google-auth==2.6.0
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.0
googleapis-common-protos==1.55.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
jsonify==0.5
MarkupSafe==2.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
